add feature (check session timeout)
